Section class extracted into the separate file
